+++ gulp/tasks/publish.js
@@ -32,20 +32,21 @@ var packages = ['paper-jsdom', 'paper-jsdom-canvas'],
         end_with_newline: true
     };
 
-gulp.task('publish', function() {
+gulp.task('publish', function(callback) {
     if (options.branch !== 'develop') {
         throw new Error('Publishing is only allowed on the develop branch.');
     }
     // publish:website comes before publish:release, so paperjs.zip file is gone
     // before npm publish:
-    return run(
+    run(
         'publish:json',
         'publish:dist',
         'publish:packages',
         'publish:commit',
         'publish:website',
         'publish:release',
-        'publish:load'
+        'publish:load',
+        callback
     );
 });
 
@@ -82,15 +83,11 @@ gulp.task('publish:release', function() {
         .pipe(shell('npm publish'));
 });
 
-gulp.task('publish:packages', function() {
-    // Publish packages in series instead of in parallel, to see if this fixes
-    // recent issues with `npm publish`:
-    return run(
-        packages.map(function(name) {
-            return 'publish:packages:' + name;
-        })
-    );
-});
+gulp.task('publish:packages',
+    packages.map(function(name) {
+        return 'publish:packages:' + name;
+    })
+);
 
 packages.forEach(function(name) {
     gulp.task('publish:packages:' + name, ['publish:version'], function() {
@@ -112,11 +109,12 @@ packages.forEach(function(name) {
     });
 });
 
-gulp.task('publish:website', function() {
+gulp.task('publish:website', function(callback) {
     if (fs.lstatSync(sitePath).isDirectory()) {
-        return run(
+        run(
             'publish:website:build',
-            'publish:website:push'
+            'publish:website:push',
+            callback
         );
     }
 });