+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 # Change Log
 
+## `0.12.3`
+
+### Fixed
+
+- Fix regression in `Color` change propagation (#1672, #1674).
+- SVG Export: Fix viewport size of exported `Symbol` (#1668).
+- Handle non-invertible matrices in `Item#contains()` (#1651).
+- Improve documentation for `Item#clipMask` (#1673).
+- Improve TypeScript definitions (#1659, #1663, #1664, #1667)
+
+### Added
+
+- Add documentation for `Item#internalBounds`.
+
 ## `0.12.2`
 
 ### Fixed
@@ -22,7 +36,7 @@
 
 ### Added
 
-- Add TypesScript definition, automatically generated from JSDoc comments
+- Add TypeScript definition, automatically generated from JSDoc comments
   (#1612).
 - Support `new Raster(size[, position])` constructor.
 - Expose `Raster#context` accessor.